backend/requirements.txt
django==4.2
psycopg2-binary==2.9.10
requests
crewai==0.1.0
Pillow
django-ckeditor

